+++ src/main/java/Task6.java
@@ -1,17 +1,17 @@
 public class Task6 {
     public static void main(String[] args) {
-        int w = 10;
-        int h = 17;
+        float w = 10;
+        float h = 17;
 
         square(w, h);
         perimetres(w, h);
     }
 
-    private static void perimetres(int w, int h) {
+    private static void perimetres(float w, float h) {
         System.out.println("Периметр прямоугольника равен " + w * h);
     }
 
-    private static void square(int w, int h) {
+    private static void square(float w, float h) {
         System.out.println("Периметр прямоугольника равен " + (w + h) * 2);
     }
 }
